requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.4.2
